LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/fetch-request-xhr.https-expected.txt

CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load https://127.0.0.1:9443/dummy?test due to access control checks.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load https://127.0.0.1:9443/dummy?test due to access control checks.

PASS initialize global state 
FAIL event.request has the expected headers for same-origin GET. promise_test: Unhandled rejection with value: object "TypeError: undefined is not an object (evaluating 'headers.length')"
FAIL event.request has the expected headers for same-origin POST. promise_test: Unhandled rejection with value: object "TypeError: undefined is not an object (evaluating 'headers.length')"
FAIL event.request has the expected headers for cross-origin GET. promise_test: Unhandled rejection with value: "XHR should succeed."
FAIL event.request has the expected headers for cross-origin POST. promise_test: Unhandled rejection with value: "XHR should succeed."
FAIL FetchEvent#request.body contains XHR request data (string) promise_test: Unhandled rejection with value: object "Error: assert_equals: expected (string) "POST" but got (undefined) undefined"
FAIL FetchEvent#request.body contains XHR request data (blob) promise_test: Unhandled rejection with value: object "Error: assert_equals: expected (string) "POST" but got (undefined) undefined"
FAIL FetchEvent#request.method is set to XHR method promise_test: Unhandled rejection with value: object "Error: assert_equals: expected (string) "XXX" but got (undefined) undefined"
FAIL XHR using OPTIONS method promise_test: Unhandled rejection with value: object "Error: assert_equals: expected (string) "OPTIONS" but got (undefined) undefined"
FAIL XHR with form data promise_test: Unhandled rejection with value: object "Error: assert_equals: expected (string) "POST" but got (undefined) undefined"
FAIL XHR with mode/credentials set promise_test: Unhandled rejection with value: object "Error: assert_equals: expected (string) "cors" but got (undefined) undefined"
PASS XHR to data URL 
PASS restore global state 
